Commit za tekucu nedelju
Dodao sam manja sredjivanja i brisanje cvorova,
Dodao sam crtanje siluete(Nisam znao da ima QSlider stavicu to nekako sam sa gradientom napravio za sad),
Dodao sam prikaz preciznosti i matrice konfuzije (trenutno samo za knn cvor posto ne radi funkcija iz nekog razlog),
Nacrtao ikonice za nove cvorove(Qt nesto ne podrzava najbolje svg fajlove strele i font se ne prikazuje uvek).

+++ Raccoon/Sources/dbscan.cpp
@@ -48,7 +48,7 @@ table dbscan::fit(const table &t)
     for(int i = 0; i<t.row_n(); i++) {
         if(train["cluster"][i] != entry(-1)) //already clustered
             continue;
-        if(eps_okolina[i].size() < n_neighbors) {
+        if(eps_okolina[i].size() < (unsigned)n_neighbors) {
             train["cluster"][i] = entry("NOISE");
             continue;
         }
@@ -59,7 +59,7 @@ table dbscan::fit(const table &t)
         for(int v: eps_okolina[i]) {
             if(train["cluster"][v] == entry("NOISE") || train["cluster"][v] == entry(-1)) //if it's classified as noise or not classified yet, it belongs to current cluster
                 train["cluster"][v] = entry(cluster_ctr);
-            if(eps_okolina[v].size() >= n_neighbors)
+            if((int)eps_okolina[v].size() >= n_neighbors)
                 eps_okolina[i].insert(eps_okolina[v].begin(), eps_okolina[v].end());
         }
     }
